Create descending_loops.c
The Descent
"Destroy the mountains before your starship collides with one of them. For that, shoot the highest mountain on your path. At the start of each game turn, you are given the height of the 8 mountains from left to right on the stdin. By the end of the game turn, you must fire on the highest mountain by outputting its index (from 0 to 7). Firing on a mountain will only destroy part of it, reducing its height. Your ship descends after each pass."

- codingame.com